WIP add third time to DoneAddingBlock event
This one forces me to leak UTCTime back into parts of Consensus. I'm very
suspicious of this.

The most attractive alternative is to exclude UTCTime from the third time in
the event. This would require the downstream event-processing logic to enrich
its tracking of times so that the event could be processed (eg rendered to the
log/as JSON/etc) knowing the time at which it was emitted.